+++ obi/task/task.py
@@ -10,6 +10,8 @@ import fabric
 import os
 import stat
 import yaml
+import re
+import pkg_resources
 
 from fabric.api import env  # the global env variable
 from fabric.api import (local, run) # the global env variable
@@ -67,6 +69,9 @@ def room_task(room_name, task_name=None):
     env.config = config_no_rooms
     env.config.update(room)
 
+    env.obi_task_name = task_name or env.tasks[-1]
+    env.obi_version = pkg_resources.require("oblong-obi")[0].version
+
     # Calling basename on project_name should be harmless
     # In the case that the user specified target, say, build/foo,
     # then basename gives us foo
@@ -77,6 +82,7 @@ def room_task(room_name, task_name=None):
     # - User specified is-local: True
     # - Room name is localhost
     # - Hosts is empty
+
     if room.get("is-local", room_name == "localhost" or not room.get("hosts", [])):
         env.hosts = ['localhost']
         env.use_ssh_config = False
@@ -84,13 +90,12 @@ def room_task(room_name, task_name=None):
         env.file_exists = os.path.exists
         env.rsync = lambda: None # Don't rsync when running locally -- noop
         env.cd = fabric.context_managers.lcd
-        # Generate a shell script that duplicates the task
-        task_name = task_name or env.tasks[-1]
         env.run = local
         env.background_run = env.run
         env.relpath = os.path.relpath
         env.launch_format_str = "{0} {1}"
         env.debug_launch_format_str = "{0} {1} {2}"
+        env.log = lambda cmd: None # no logging for local
     else:
         env.user = room.get("user", env.local_user) # needed for remote run
         env.hosts = room.get("hosts", [])
@@ -109,6 +114,7 @@ def room_task(room_name, task_name=None):
         env.relpath = lambda p: p
         env.launch_format_str = "sh -c '(({0} nohup {1} > {2} 2> {2}) &)'"
         env.debug_launch_format_str = "tmux new -d -s {0} '{1}'".format(env.target_name, "{0} {1} {2}")
+        env.log = write_to_journal
     env.build_dir = os.path.abspath(env.relpath(os.path.join(env.project_dir, env.config.get("build-dir", "build"))))
 
 @task
@@ -121,6 +127,7 @@ def build_task():
         user_specified_build = env.config.get("build-cmd", None)
         if env.config.has_key("build-cmd"):
             if user_specified_build:
+                env.log(user_specified_build)
                 env.run(user_specified_build)
         else:
             # Arguments for the cmake step
@@ -129,8 +136,11 @@ def build_task():
             # Arguments for the build step
             build_args = env.config.get("build-args", [])
             build_args = " ".join(build_args)
+            env.log("mkdir -p {0}".format(env.build_dir))
             env.run("mkdir -p {0}".format(env.build_dir))
+            env.log("cmake -H\"{0}\" -B\"{1}\" {2}".format(env.project_dir, env.build_dir, cmake_args))
             env.run("cmake -H\"{0}\" -B\"{1}\" {2}".format(env.project_dir, env.build_dir, cmake_args))
+            env.log("cmake --build \"{0}\" -- {1}".format(env.build_dir, build_args))
             env.run("cmake --build \"{0}\" -- {1}".format(env.build_dir, build_args))
 
 @task
@@ -143,9 +153,11 @@ def clean_task():
         user_specified_clean = env.config.get("clean-cmd", None)
         if env.config.has_key("clean-cmd"):
             if user_specified_clean:
+                env.log(user_specified_clean)
                 env.run(user_specified_clean)
         else:
             clean_cmd = "rm -rf {0} || true".format(env.build_dir)
+            env.log(clean_cmd)
             env.run(clean_cmd)
 
 @task
@@ -158,15 +170,18 @@ def stop_task():
         # fall-back to on-stop-cmds for backwards compatibility
         # TODO(jshrake): remove support for the amiguous on-stop-cmds key
         for cmd in env.config.get("pre-stop-cmds", env.config.get("on-stop-cmds", [])):
+            env.log(cmd)
             env.run(cmd)
     with env.cd(env.project_dir):
         for cmd in env.config.get("local-pre-stop-cmds", []):
             local(cmd)
     default_stop = "pkill -SIGTERM -f '[a-z/]+{0} .*' || true".format(env.target_name)
     stop_cmd = env.config.get("stop-cmd", default_stop)
+    env.log(stop_cmd)
     env.run(stop_cmd)
     with env.cd(env.project_dir):
         for cmd in env.config.get("post-stop-cmds", []):
+            env.log(cmd)
             env.run(cmd)
     with env.cd(env.project_dir):
         for cmd in env.config.get("local-post-stop-cmds", []):
@@ -225,6 +240,7 @@ def launch_task(debugger, extras):
     with env.cd(env.project_dir):
         # Process pre-launch commands
         for cmd in env.config.get("pre-launch-cmds", []):
+            env.log(cmd)
             env.run(cmd)
         if debugger:
             debug_cmd = debugger
@@ -233,14 +249,17 @@ def launch_task(debugger, extras):
                 debug_cmd = debuggers.get(debugger, debug_cmd)
             default_launch = env.debug_launch_format_str.format(env_vars, debug_cmd, formatted_launch)
             launch_cmd = env.config.get("debug-launch-cmd", default_launch)
+            env.log(launch_cmd)
             env.background_run(launch_cmd)
         else:
             log_file = env.relpath(os.path.join(env.project_dir, env.target_name + ".log"))
             default_launch = env.launch_format_str.format(env_vars, formatted_launch, log_file)
             launch_cmd = env.config.get("launch-cmd", default_launch)
+            env.log(launch_cmd)
             env.background_run(launch_cmd)
         # Process the post-launch commands
         for cmd in env.config.get("post-launch-cmds", []):
+            env.log(cmd)
             env.run(cmd)
 
 @task
@@ -264,7 +283,9 @@ def rsync_task():
     a new directory /home/username/foldername (if needed) and place the
     files there.
     """
+    env.log("mkdir -p {0}".format(env.project_dir))
     env.run("mkdir -p {0}".format(env.project_dir))
+    env.log("fabric rsync local={} remote={}".format(env.local_project_dir + "/", env.project_dir))
     return fabric.contrib.project.rsync_project(
         local_dir=env.local_project_dir + "/",
         remote_dir=env.project_dir,
@@ -309,3 +330,23 @@ def project_yaml():
             current = parent
             parent = parent_dir(current)
     abort("Could not find the project.yaml file in {0} or any parent directories".format(os.getcwd()))
+
+# taken from https://github.com/python/cpython/blob/c80b0175c88be9611b6eea7a60104b4488839a04/Lib/shlex.py#L308
+#_find_unsafe = re.compile(r'[^\w@%+=:,./-]', re.ASCII).search
+_find_unsafe = re.compile(r'[^\w@%+=:,./-]').search #re.ASCII is py3
+def shlexquote(s):
+    """Return a shell-escaped version of the string *s*."""
+    if not s: return "''"
+    if _find_unsafe(s) is None: return s
+
+    # use single quotes, and put single quotes into double quotes
+    # the string $'b is then quoted as '$'"'"'b'
+    return "'" + s.replace("'", "'\"'\"'") + "'"
+
+def write_to_journal(string_input):
+    """
+    Run systemd-cat echo 'obi go by arice: message' on remote hosts
+    """
+    logmessage = 'obi {} by {}: {}'.format(env.obi_task_name, env.local_user, string_input)
+    shellstring = 'if [ -x "$(command -v systemd-cat)" ]; then systemd-cat -t obi echo {}; fi'.format(shlexquote(logmessage))
+    env.run(shellstring)